Data.Image

Description

The Univeristy of New Mexico's Haskell Image Processing library contains functionality for performing manipulation gray, color, and complex images.

Documentation

By importing Data.Image you will get functionality for working with Boxed images

module Data.Image.Boxed