Data.Image.Binary

Binary Images

class BinaryPixel px where

A BinaryPixel can be in one of two states, on or off.

Methods

toBinary :: px -> Bool

off :: px

on :: px

toBinaryImage :: (Image img, BinaryPixel (Pixel img)) => (Pixel img -> Bool) -> img -> img

Given a function of a pixel to a boolean and an image, returns the Binary version of that image.

(<.) :: (Image img, BinaryPixel (Pixel img), Ord (Pixel img), Pixel img ~ a) => a -> img -> img

Given a Pixel p and an image img, return a Binary image where the pixel at (i, j) is on if p is less than the corresponding pixel in img at (i,j) and off otherwise.

(.<) :: (Image img, BinaryPixel (Pixel img), Ord (Pixel img), Pixel img ~ a) => img -> a -> img

Given an image, img, and a Pixel p, return a Binary image where the pixel at (i, j) is on if the corresponding pixel in img at (i,j) is less than p and off otherwise.

(>.) :: (Image img, BinaryPixel (Pixel img), Ord (Pixel img), Pixel img ~ a) => a -> img -> img

Given a Pixel p and an image img, return a Binary image where the pixel at (i, j) is on if p is greater than the corresponding pixel in img at (i,j) and off otherwise.

(.>) :: (Image img, BinaryPixel (Pixel img), Ord (Pixel img), Pixel img ~ a) => img -> a -> img

Given an image, img, and a Pixel p, return a Binary image where the pixel at (i, j) is on if the corresponding pixel in img at (i,j) is greater than p and off otherwise.

(==.) :: (Image img, BinaryPixel (Pixel img), Eq (Pixel img), Pixel img ~ a) => a -> img -> img

Given a Pixel p and an image img, return a Binary image where the pixel at (i, j) is on if the corresponding pixel in img at (i,j) is equal to p and off otherwise.

(.==) :: (Image img, BinaryPixel (Pixel img), Eq (Pixel img), Pixel img ~ a) => img -> a -> img

Given an image, img, and a Pixel p, return a Binary image where the pixel at (i, j) is on if the corresponding pixel in img at (i,j) is equal to p and off otherwise.

compareImage :: (Image img, BinaryPixel (Pixel img), Ord (Pixel img)) => (Pixel img -> Pixel img -> Bool) -> img -> img -> img

Given a function of two pixels to a boolean pred and two images X and Y, return a binary image that for each pixel (i,j) is on if the pixel pred X(i,j) Y(i,j) return True and off otherwise.

(.<.) :: (Image img, BinaryPixel (Pixel img), Ord (Pixel img)) => img -> img -> img

Given two images X and Y, return a binary image that for each pixel (i, j) is on if X(i,j) < Y(i,j) and off otherwise.

(.>.) :: (Image img, BinaryPixel (Pixel img), Ord (Pixel img)) => img -> img -> img

Given two images X and Y, return a binary image that for each pixel (i, j) is on if X(i,j) > Y(i,j) and off otherwise.

(.==.) :: (Image img, BinaryPixel (Pixel img), Eq (Pixel img)) => img -> img -> img

Given two images X and Y, return a bniry image that for each pixel (i, j) is on if X(i,j) == Y(i,j) and off otherwise.

Binary Morphology

erode :: (Image img, BinaryPixel (Pixel img), Num (Pixel img), Eq (Pixel img)) => [[Pixel img]] -> img -> img

Given a 2D list consisting solely of pixels representing a structuring element, and a binary image, erode returns the morphological erosion of the image with the structuring element.

erode' :: (Image img, BinaryPixel (Pixel img), Num (Pixel img), Eq (Pixel img)) => img -> img

For convenience erode' = erode [[1,1],[1,1]]

dilate :: (Image img, BinaryPixel (Pixel img), Num (Pixel img), Ord (Pixel img)) => [[Pixel img]] -> img -> img

Given a 2D list consisting solely of pixels representing a structuring element, and a binary image, dilate returns the morphological dilation of the image with the structuring element.

dilate' :: (Image img, BinaryPixel (Pixel img), Num (Pixel img), Ord (Pixel img)) => img -> img

For convenience dilate' = dilate [[1,1],[1,1]]

open :: (Image img, BinaryPixel (Pixel img), Num (Pixel img), Ord (Pixel img)) => [[Pixel img]] -> img -> img

Given a 2D list consisting solely of pixels representing a structuring element, and a binary image, dilate returns the morphological opening of the image with the structuring element.

open' :: (Image img, BinaryPixel (Pixel img), Num (Pixel img), Ord (Pixel img)) => img -> img

For convenience open' = open [[1,1],[1,1]]

close :: (Image img, BinaryPixel (Pixel img), Num (Pixel img), Ord (Pixel img)) => [[Pixel img]] -> img -> img

Given a 2D list consisting solely of pixels representing a structuring element, and a binary image, dilate returns the morphological closing of the image with the structuring element.

close' :: (Image img, BinaryPixel (Pixel img), Num (Pixel img), Ord (Pixel img)) => img -> img

For convenience close' = close [[1,1],[1,1]]

Functions on Binary Images

label :: (Image img, BinaryPixel (Pixel img), Image img', Pixel img' ~ Double) => img -> img'

Given a binary image, label returns an image where pixels in distinct connected components (based on 4-neighbor connectivity) have distinct integer values. These values range from 1 to n where n is the number of connected components in image.

areas :: (Image img, MaxMin (Pixel img), RealFrac (Pixel img)) => img -> Vector Double

Given an image, areas returns a vector where the n-th component equals the number of pixels with value n. If image is the result of applying label to a binary image, then the vector represents the areas of the connected-components of the binary-image. If not, areas returns the histogram of the image.

perimeters :: (Image img, MaxMin (Pixel img), Pixel img ~ Double) => img -> Vector Double

Given an image, perimeters returns a vector where the n-th component equals the number of pixels with value n which are adjacent to pixels of value 0 and the 0-th component equals the sum of the other components. If image is the result of applying label to a binary image, then the vector represents the perimeters of the connected-components of the binary-image.

boundingBoxes :: (Image img, MaxMin (Pixel img), Pixel img ~ Double) => img -> [(Int, Int, Int, Int)]

Given an image, the result of applying label to a binary-image, boundingBoxes returns a vector where the n-th component is a four element tuple representing the minimum and maximum row and column indices of pixels of the n-th connected-component of the image.

centersOfMass :: (Image img, MaxMin (Pixel img), Pixel img ~ Double) => img -> [(Double, Double)]

Given an image, the result of applying label to a binary-image, centersOfMass returns a vector where the n-th component is a tuple representing the average row and column indices of pixels of the n-th connected-component of the image.

distanceTransform :: (Image img, BinaryPixel (Pixel img), Image img', Pixel img' ~ Double) => img -> img'

Given a binary image, distanceTransform returns an image representing the 2D distance transform of the image. The distance transform is accurate to within a 2% error for euclidean distance.

outline :: (Image img, BinaryPixel (Pixel img), Image img', Pixel img' ~ Double) => img -> img'

Given an image, outline returns an image where edge pixels are set to the value 1 and non-edge pixels are set to the value 0. Pixel (i, j) is an edge pixel iff its value is different than the value of either pixel (i, j+1) or pixel (i+1, j).

outline' :: (Image img, BinaryPixel (Pixel img), Image img', Pixel img' ~ Double) => Double -> Double -> img -> img'

Given two doubles nonEdge and edge, and an image, outline' returns an image where edge pixels are set to the value edge and non-edge pixels are set to the value nonEdge. Pixel (i, j) is an edge pixel iff its value is different than the value of either pixel (i, j+1) or pixel (i+1, j).