Data.Image.IO

Documentation

class DisplayFormat df where

A DisplayFormat for writing to a file

Methods

format :: df -> String

class RealFrac (GrayVal px) => GrayPixel px where

GrayPixels will be converted using this class

Associated Types

type GrayVal px :: *

Methods

toGray :: px -> GrayVal px

class RealFrac (ColorVal px) => RGBPixel px where

RGBPixels will be converted using this class

Associated Types

type ColorVal px :: *

Methods

toRGB :: px -> (ColorVal px, ColorVal px, ColorVal px)

writeImage :: DisplayFormat df => FilePath -> df -> IO ()

Given a file name and a formatable image, writes the image to that file with the format.

toPGM :: (Image img, GrayPixel (Pixel img)) => img -> [Char]

Converts an image an ASCII PPM scaled between pixel values of 0 and 255

toPPM :: (Image img, RGBPixel (Pixel img)) => img -> [Char]

Converts an image to an ASCII PPM scaled between pixel values of 0 and 255