Data.Image.Boxed

Documentation

data BoxedImage a

BoxedImage is a concrete implementation of Image using a boxed internal structure. This allows for it to be installed nicely in Functor and Applicative.

Gray Images

type GrayImage = BoxedImage Gray

A concrete instance of Image representing a gray scale image. This instance is installed in DisplayFormat as a gray PGM.

type Gray = Double

readImage :: FilePath -> IO GrayImage

Reads in a ASCII PGM image located at fileName as a GrayImage

grayToComplex :: GrayImage -> ComplexImage

Coerces a GrayImage to a ComplexImage where the imaginary part for all pixels is 0.

makeHotImage :: GrayImage -> ColorImage

Given a GrayImage, makeHotImage returns a ColorImage with the same dimensions. The R, G, B values of the result image at (i, j) are determined by using the value of the ColorImage at (i, j) to index three lookup tables. These lookup tables implement a false coloring scheme which maps small values to black, large values to white, and intermediate values to shades of red, orange, and yellow (in that order).

Color Images

type ColorImage = BoxedImage Color

A concrete instance of Image that represents images with color values. This instance is installed in DisplayFormat and can be written to a color PPM

data Color

A color encoding scheme

Constructors

RGB (Double, Double, Double)
Red, Green, Blue encoding

HSI (Double, Double, Double)
Hue, Saturation, Intensity encoding

readColorImage :: FilePath -> IO ColorImage

Reads in an ASCI PPM file as a ColorImage

colorImageRed :: ColorImage -> GrayImage

Given a ColorImage, returns a GrayImage representing the Red color component

colorImageGreen :: ColorImage -> GrayImage

Given a ColorImage, returns a GrayImage representing the Green color component

colorImageBlue :: ColorImage -> GrayImage

Given a ColorImage, returns a GrayImage representing the Blue color component

colorImageToRGB :: ColorImage -> (GrayImage, GrayImage, GrayImage)

Given a ColorImage, returns a triple containing three GrayImages each containing one of the color components (red, green, blue)

rgbToColorImage :: (GrayImage, GrayImage, GrayImage) -> ColorImage

Given a triple containing three GrayImages each containing one of the color components (red, green, blue), returns a ColorImage

colorImageHue :: ColorImage -> GrayImage

Given a ColorImage, returns a GrayImage representing the Hue component

colorImageSaturation :: ColorImage -> GrayImage

Given a ColorImage, returns a GrayImage representing the Saturation component

colorImageIntensity :: ColorImage -> GrayImage

Given a ColorImage, returns a GrayImage representing the Intensity component

colorImageToHSI :: ColorImage -> (GrayImage, GrayImage, GrayImage)

Given a triple containing three GrayImages each containing one of the color components (hue, saturation, ), returns a ColorImage

hsiToColorImage :: (GrayImage, GrayImage, GrayImage) -> ColorImage

Given a ColorImage, returns a triple containing three GrayImages each containing one of the components (hue, saturation, intensity)

Complex Images

type ComplexImage = BoxedImage Complex

A concrete instance of Image representing pixels as complex values. This instance can be written to file as a color PPM.

type Complex = Complex Double

realPart :: ComplexImage -> GrayImage

Given a complex image, returns a real image representing the real part of the image.

imagPart :: ComplexImage -> GrayImage

Given a complex image, returns a real image representing the imaginary part of the image

magnitude :: ComplexImage -> GrayImage

Given a complex image, returns a real image representing the magnitude of the image.

angle :: ComplexImage -> GrayImage

Given a complex image, returns a real image representing the angle of the image

complex :: GrayImage -> GrayImage -> ComplexImage

Given an image representing the real part of a complex image, and an image representing the imaginary part of a complex image, returns a complex image.

complexImageToRectangular :: ComplexImage -> (GrayImage, GrayImage)

Given a complex image, return a pair of real images each representing a component of the complex image (real, imaginary).

shrink :: (Num a, Image img, ComplexPixel (Pixel img), Value (Pixel img) ~ Double) => a -> img -> ComplexImage

Given a complex image and a real positive number x, shrink returns a complex image with the same dimensions. Let z be the value of the image at location (i, j). The value of the complex result image at location (i, j) is zero if |z| < x, otherwise the result has the same phase as z but the amplitude is decreased by x.

makeFilter :: Image img => Int -> Int -> PixelOp (Pixel img) -> img

Given a positive integer m, a positive integer n, and a function returning a pixel value, makeFilter returns an image with m rows and n columns. Let x equal i if i is less than m/2 and i - m otherwise and let y equal j if j is less than n/2 and j - n otherwise. To match the periodicity of the 2D discrete Fourier spectrum, the value of the result image at location (i, j) is computed by applying the function to x and y, e.g., the value at location (0, 0) is the result of applying the function to 0 and 0, the value at (m-1, n-1) is the result of applying function to -1 and -1.

fft :: (Image img, ComplexPixel (Pixel img), Value (Pixel img) ~ Double) => img -> ComplexImage

Given an image whose pixels can be converted to a complex value, fft returns an image with complex pixels representing its 2D discrete Fourier transform (DFT). Because the DFT is computed using the Fast Fourier Transform (FFT) algorithm, the number of rows and columns of the image must both be powers of two, i.e., 2K where K is an integer.

ifft :: (Image img, ComplexPixel (Pixel img), Value (Pixel img) ~ Double) => img -> ComplexImage

Given an image, ifft returns a complex image representing its 2D inverse discrete Fourier transform (DFT). Because the inverse DFT is computed using the Fast Fourier Transform (FFT) algorithm, the number of rows and columns of image must both be powers of two, i.e., 2K where K is an integer.

Binary Images

distanceTransform :: (Image img, BinaryPixel (Pixel img)) => img -> GrayImage

Given a binary image, distanceTransform returns an image representing the 2D distance transform of the image. The distance transform is accurate to within a 2% error for euclidean distance.

label :: (Image img, BinaryPixel (Pixel img)) => img -> GrayImage

Given a binary image, label returns an image where pixels in distinct connected components (based on 4-neighbor connectivity) have distinct integer values. These values range from 1 to n where n is the number of connected components in image.

Additional Modules

Contains functionality related to Binary Images

module Data.Image.Binary

Contains functionality for convolution of images

module Data.Image.Convolution

Contains basic functionality for Images

module Data.Image.Internal

Contains functionality for writing images and displaying with an external program

module Data.Image.IO